refactor : List<LottoTicket> 을 감싸는 LottoTickets 객체 추가
