+++ src/main/java/lotto3/controller/LottoController.java
@@ -2,10 +2,9 @@ package lotto3.controller;
 
 import java.util.List;
 import java.util.Map;
-import lotto3.domain.LottoTicket;
+import lotto3.domain.LottoTickets;
 import lotto3.domain.Prize;
 import lotto3.util.LottoFactory;
-import lotto3.util.LottoResultCalculator;
 import lotto3.view.InputView;
 import lotto3.view.OutputView;
 
@@ -13,12 +12,13 @@ public class LottoController {
 
   public static void main(String[] args) {
     int investMoney = InputView.scanMoney();
-    List<LottoTicket> tickets = LottoFactory.createLottoTickets(investMoney);
+    LottoTickets tickets = LottoFactory.createLottoTickets(investMoney);
 
     OutputView.printLottoTickets(tickets);
 
     List<Integer> winningNumbers = InputView.scanWinningNumbers();
-    Map<Prize, Long> lottoResults = LottoResultCalculator.calculate(tickets, winningNumbers);
+
+    Map<Prize, Long> lottoResults = tickets.calculateLotteryResults(winningNumbers);
 
     OutputView.printLottoResults(lottoResults);
     OutputView.printProfitRate(lottoResults, investMoney);
